subpixel tracking and fixed fit matrix

+++ test_app/client_lib/cuda_transfer.h
@@ -17,7 +17,7 @@ LIB_FN void beamform_i16( const short* data, CudaBeamformerParameters bp, float*
 
 LIB_FN void beamform_f32(const float* data, CudaBeamformerParameters bp, float* output);
 
-LIB_FN void motion_detect_f32(const float* images, NCCMotionParameters params, int* motion_maps);
+LIB_FN void motion_detect_f32(const float* images, NCCMotionParameters params, float* motion_maps);
 
 LIB_FN void deinit();
